requirements.txt
altgraph==0.17.4
anyio==4.6.0
APScheduler==3.10.4
certifi==2024.8.30
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
psutil==6.0.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-telegram-bot==21.6
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wxPython==4.2.2
